requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
phonenumbers==9.0.10
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2

